bots/cai-bot/requirements.txt
characterai
bottom==2.2.0
trio-asyncio==0.12.0
tls-client==1.0.1
typing-extensions==4.10.0
python-dotenv==1.0.1
websockets==12.0
beautifulsoup4==4.12.3
async-lru==2.0.4
